refactor(playground): split template gallery data
