+++ packages/agent-playground/src/components/playground/template-gallery-data/index.ts
@@ -1,0 +1,3 @@
+export { categoryIcons, difficultyColors, providerIcons } from './display';
+export { templates } from './templates';
+export type { ITemplate, ITemplateGalleryProps } from './types';
